Separate benefits into comp and add to clinic page
